Update CleanCSV.py
The purpose of this script is to change the CSV into a pipe-delimited file. The issue is that the data may contain pipe characters and this script does not take that into account. There are also issues with the unknown characters due to an unknown character set provided in the file.  All character sets were tested and they all threw errors. Instead the open() command has an ignore errors parameter passed and this works fine.